// src/api/model/ticketStatus.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * e621 API
 * An API for accessing user information and other resources on e621 and e926.
 * OpenAPI spec version: 1.0.0
 */

/**
 * The current status of the ticket
 */
export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TicketStatus = {
  pending: "pending",
  approved: "approved",
  partial: "partial",
} as const;